+++ packages/axios-extension/src/abap/adt-catalog/services/index.ts
@@ -5,5 +5,4 @@ export { TransportRequestService, NewUi5ObjectRequestParams } from './transportr
 export { ListPackageService } from './list-package-service';
 export { FileStoreService } from './filestore-service';
 export { GeneratorService } from './generator-service';
-export { PublishService } from './publish-service';
 export { BusinessObjectsService } from './businessobjects-service';