+++ packages/preview-middleware-client/src/adp/init-dialogs.ts
@@ -30,33 +30,6 @@ export const enum DialogNames {
 type Controller = AddFragment | ControllerExtension | ExtensionPoint;
 
 /**
- * Adds a new item to the context menu
- *
- * @param rta Runtime Authoring
- * @param syncViewsIds Ids of all application sync views
- * @param minorUI5Version minor UI5 version
- */
-export const initDialogs = (rta: RuntimeAuthoring, syncViewsIds: string[], minorUI5Version: number): void => {
-    const contextMenu = rta.getDefaultPlugins().contextMenu;
-
-    contextMenu.addMenuItem({
-        id: 'ADD_FRAGMENT',
-        text: getAddFragmentItemText,
-        handler: async (overlays: UI5Element[]) => await handler(overlays[0], rta, DialogNames.ADD_FRAGMENT),
-        icon: 'sap-icon://attachment-html',
-        enabled: (overlays: ElementOverlay[]) => isFragmentCommandEnabled(overlays, minorUI5Version)
-    });
-
-    contextMenu.addMenuItem({
-        id: 'EXTEND_CONTROLLER',
-        text: 'Extend With Controller',
-        handler: async (overlays: UI5Element[]) => await handler(overlays[0], rta, DialogNames.CONTROLLER_EXTENSION),
-        icon: 'sap-icon://create-form',
-        enabled: (overlays: ElementOverlay[]) => isControllerExtensionEnabled(overlays, syncViewsIds, minorUI5Version)
-    });
-};
-
-/**
  * Handler for enablement of Extend With Controller context menu entry
  *
  * @param overlays Control overlays
@@ -157,3 +130,30 @@ export async function handler(
 
     await controller.setup(dialog as Dialog);
 }
+
+/**
+ * Adds a new item to the context menu
+ *
+ * @param rta Runtime Authoring
+ * @param syncViewsIds Ids of all application sync views
+ * @param minorUI5Version minor UI5 version
+ */
+export const initDialogs = (rta: RuntimeAuthoring, syncViewsIds: string[], minorUI5Version: number): void => {
+    const contextMenu = rta.getDefaultPlugins().contextMenu;
+
+    contextMenu.addMenuItem({
+        id: 'ADD_FRAGMENT',
+        text: getAddFragmentItemText,
+        handler: async (overlays: UI5Element[]) => await handler(overlays[0], rta, DialogNames.ADD_FRAGMENT),
+        icon: 'sap-icon://attachment-html',
+        enabled: (overlays: ElementOverlay[]) => isFragmentCommandEnabled(overlays, minorUI5Version)
+    });
+
+    contextMenu.addMenuItem({
+        id: 'EXTEND_CONTROLLER',
+        text: 'Extend With Controller',
+        handler: async (overlays: UI5Element[]) => await handler(overlays[0], rta, DialogNames.CONTROLLER_EXTENSION),
+        icon: 'sap-icon://create-form',
+        enabled: (overlays: ElementOverlay[]) => isControllerExtensionEnabled(overlays, syncViewsIds, minorUI5Version)
+    });
+};