+++ packages/preview-middleware-client/src/cpe/outline/index.ts
@@ -6,6 +6,7 @@ import type OutlineService from 'sap/ui/rta/command/OutlineService';
 
 import { transformNodes } from './nodes';
 import Log from 'sap/base/Log';
+import { getError } from '../error-utils';
 
 /**
  *
@@ -28,7 +29,7 @@ export async function initOutline(rta: RuntimeAuthoring, sendAction: (action: Ex
             } 
            
         } catch (error) {
-            Log.error('Outline sync failed!', error);
+            Log.error('Outline sync failed!', getError(error));
         }
     }
     await syncOutline();