+++ packages/control-property-editor-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux-private/control-property-editor-common
 
+## 0.4.0
+
+### Minor Changes
+
+-   c2359077: [BREAKING CHANGE] Change TypeScript transpile target to ES2021 to align with NodeJS 18+
+
 ## 0.3.6
 
 ### Patch Changes